+++ frontend/gulp/configs/reduce-css.json
@@ -8,6 +8,15 @@
         "resources/components/header/*.ejs"
       ],
       "out": "resources/pages/homepage/views/used.scss"
+    },
+    {
+      "path": [
+        "resources/pages/pdp/**/*.ejs",
+        "resources/components/master/*.ejs",
+        "resources/components/footer/*.ejs",
+        "resources/components/header/*.ejs"
+      ],
+      "out": "resources/pages/pdp/views/used.scss"
     }
   ],
   "baseCSS": [
