+++ frontend/resources/pages/homepage/views/homepage.ejs
@@ -10,10 +10,6 @@
     <%-     include("./components/content-adp/view.ejs") -%>
 </div>
 
-<div>
-    <%-     include("./components/atf-pdp/view.ejs") -%>
-</div>
-
 <div class="pa3">
     <%-     include("./components/products-list/view.ejs") -%>
 </div>
@@ -34,13 +30,6 @@
 <div class="pa3">
     <%-     include("./components/top-categories/view.ejs") -%>
 </div>
-<div>
-    <%-     include("./components/related-articles/view.ejs") -%>
-</div>
-<div>
-    <%-     include("./components/content-pdp/view.ejs") -%>
-</div>
-
 <div class="pa3">
     <%-     include("./components/categories-sidebar/view.ejs") -%>
 </div>
