+++ frontend/resources/pages/pdp/data-feed/pdp.json
@@ -1,0 +1,10 @@
+{
+  "title": "Hai Bui Cool",
+  "heading": "Hai Bui oi co len",
+  "newProducts": [],
+  "variantGroups": [],
+  "variant": {},
+  "productDetail": {},
+  "css": "",
+  "canonical": "/"
+}
