+++ frontend/resources/pages/pdp/route.js
@@ -1,0 +1,6 @@
+var router = require('express').Router();
+const controller = require("./pdpController");
+
+router.get('/san-pham/:productID/', controller);
+
+module.exports = router;
