+++ frontend/resources/pages/pdp/views/components/atf-pdp/product-images.ejs
