+++ frontend/resources/pages/pdp/views/components/atf-pdp/variant.ejs
