+++ frontend/resources/pages/pdp/views/components/atf-pdp/variants.ejs
