+++ frontend/resources/pages/pdp/views/components/atf-pdp/view.ejs
