+++ frontend/resources/pages/pdp/views/components/content-pdp/view.ejs
