+++ frontend/resources/pages/pdp/views/components/related-articles/view.ejs
