+++ frontend/resources/pages/pdp/views/ifarmer-pdp.scss
@@ -1,0 +1,2 @@
+@import "../../../styles/standards/index";
+@import "used";
