+++ frontend/resources/pages/pdp/views/pdp.ejs
@@ -1,0 +1,19 @@
+<%- include("../../../components/master/head") -%>
+
+<h1 class="f2 f1-l"><%= heading %></h1>
+
+<div>
+    <%-     include("./components/atf-pdp/view.ejs") -%>
+</div>
+
+<div>
+    <%-     include("./components/related-articles/view.ejs") -%>
+</div>
+
+<div>
+    <%-     include("./components/content-pdp/view.ejs") -%>
+</div>
+
+
+
+<%- include("../../../components/master/footer") -%>
