+++ frontend/src/variant-types/variantTypesService.js
@@ -1,4 +1,4 @@
-const {search} = require("../../../../src/elasticsearch/search");
+const {search} = require("../elasticsearch/search");
 
 const searchVariantTypes = search("variant-types");
 const getAllVariantTypes = async () => {
